+++ examples/kubernetes/cluster_addon_v2/main-datasource.tf
@@ -1,0 +1,3 @@
+data "vkcs_kubernetes_cluster_addon_v2" "ingress_nginx" {
+  id = vkcs_kubernetes_cluster_addon_v2.ingress_nginx.id
+}