+++ vkcs/kubernetes/resource_cluster_v2_test.go
@@ -12,6 +12,7 @@ import (
 	"github.com/vk-cs/terraform-provider-vkcs/vkcs/internal/acctest"
 	"github.com/vk-cs/terraform-provider-vkcs/vkcs/internal/clients"
 	"github.com/vk-cs/terraform-provider-vkcs/vkcs/internal/services/kubernetes/containerinfra/v2/clusters"
+	"github.com/vk-cs/terraform-provider-vkcs/vkcs/internal/util/errutil"
 )
 
 func TestAccKubernetesClusterV2_basic(t *testing.T) {
@@ -24,47 +25,47 @@ func TestAccKubernetesClusterV2_basic(t *testing.T) {
 		ProtoV6ProviderFactories: acctest.AccTestProtoV6ProviderFactories,
 		Steps: []resource.TestStep{
 			{
-				Config: acctest.AccTestRenderConfig(testAccKubernetesClusterV2Basic, map[string]string{
+				Config: acctest.AccTestRenderConfig(testAccKubernetesClusterV2Base, map[string]string{
 					"TestAccKubernetesNetworkingBase": testAccKubernetesNetworkingBase,
 					"ClusterName":                     clusterName,
 				}),
 				Check: resource.ComposeTestCheckFunc(
-					testAccCheckKubernetesClusterV2Exists("vkcs_kubernetes_cluster_v2.basic"),
-					resource.TestCheckResourceAttr("vkcs_kubernetes_cluster_v2.basic", "name", clusterName),
-					resource.TestCheckResourceAttr("vkcs_kubernetes_cluster_v2.basic", "version", "v1.34.2"),
-					resource.TestCheckResourceAttr("vkcs_kubernetes_cluster_v2.basic", "cluster_type", "standard"),
-					resource.TestCheckResourceAttr("vkcs_kubernetes_cluster_v2.basic", "availability_zones.#", "1"),
-					resource.TestCheckResourceAttr("vkcs_kubernetes_cluster_v2.basic", "availability_zones.0", "MS1"),
-					resource.TestCheckResourceAttr("vkcs_kubernetes_cluster_v2.basic", "master_count", "1"),
-					resource.TestCheckResourceAttr("vkcs_kubernetes_cluster_v2.basic", "network_plugin", "calico"),
-					resource.TestCheckResourceAttr("vkcs_kubernetes_cluster_v2.basic", "pods_ipv4_cidr", "10.100.0.0/16"),
-					resource.TestCheckResourceAttr("vkcs_kubernetes_cluster_v2.basic", "public_ip", "false"),
-					resource.TestCheckResourceAttr("vkcs_kubernetes_cluster_v2.basic", "master_disks.#", "3"),
-					resource.TestCheckResourceAttr("vkcs_kubernetes_cluster_v2.basic", "description", ""),
-					resource.TestCheckResourceAttr("vkcs_kubernetes_cluster_v2.basic", "api_lb_fip", ""),
-					resource.TestCheckResourceAttr("vkcs_kubernetes_cluster_v2.basic", "insecure_registries.#", "0"),
-					resource.TestCheckResourceAttr("vkcs_kubernetes_cluster_v2.basic", "labels.#", "0"),
-					resource.TestCheckResourceAttr("vkcs_kubernetes_cluster_v2.basic", "loadbalancer_allowed_cidrs.#", "0"),
-					resource.TestCheckResourceAttr("vkcs_kubernetes_cluster_v2.basic", "node_groups.#", "0"),
+					testAccCheckKubernetesClusterV2Exists("vkcs_kubernetes_cluster_v2.base"),
+					resource.TestCheckResourceAttr("vkcs_kubernetes_cluster_v2.base", "name", clusterName),
+					resource.TestCheckResourceAttr("vkcs_kubernetes_cluster_v2.base", "version", "v1.34.2"),
+					resource.TestCheckResourceAttr("vkcs_kubernetes_cluster_v2.base", "cluster_type", "standard"),
+					resource.TestCheckResourceAttr("vkcs_kubernetes_cluster_v2.base", "availability_zones.#", "1"),
+					resource.TestCheckResourceAttr("vkcs_kubernetes_cluster_v2.base", "availability_zones.0", "MS1"),
+					resource.TestCheckResourceAttr("vkcs_kubernetes_cluster_v2.base", "master_count", "1"),
+					resource.TestCheckResourceAttr("vkcs_kubernetes_cluster_v2.base", "network_plugin", "calico"),
+					resource.TestCheckResourceAttr("vkcs_kubernetes_cluster_v2.base", "pods_ipv4_cidr", "10.100.0.0/16"),
+					resource.TestCheckResourceAttr("vkcs_kubernetes_cluster_v2.base", "public_ip", "false"),
+					resource.TestCheckResourceAttr("vkcs_kubernetes_cluster_v2.base", "master_disks.#", "3"),
+					resource.TestCheckResourceAttr("vkcs_kubernetes_cluster_v2.base", "description", ""),
+					resource.TestCheckResourceAttr("vkcs_kubernetes_cluster_v2.base", "api_lb_fip", ""),
+					resource.TestCheckResourceAttr("vkcs_kubernetes_cluster_v2.base", "insecure_registries.#", "0"),
+					resource.TestCheckResourceAttr("vkcs_kubernetes_cluster_v2.base", "labels.#", "0"),
+					resource.TestCheckResourceAttr("vkcs_kubernetes_cluster_v2.base", "loadbalancer_allowed_cidrs.#", "0"),
+					resource.TestCheckResourceAttr("vkcs_kubernetes_cluster_v2.base", "node_groups.#", "0"),
 
-					resource.TestCheckResourceAttrPair("vkcs_kubernetes_cluster_v2.basic", "master_flavor", "data.vkcs_compute_flavor.base", "id"),
-					resource.TestCheckResourceAttrPair("vkcs_kubernetes_cluster_v2.basic", "network_id", "vkcs_networking_network.base", "id"),
-					resource.TestCheckResourceAttrPair("vkcs_kubernetes_cluster_v2.basic", "subnet_id", "vkcs_networking_subnet.base", "id"),
-					resource.TestCheckResourceAttrPair("vkcs_kubernetes_cluster_v2.basic", "loadbalancer_subnet_id", "vkcs_networking_subnet.base", "id"),
+					resource.TestCheckResourceAttrPair("vkcs_kubernetes_cluster_v2.base", "master_flavor", "data.vkcs_compute_flavor.master", "id"),
+					resource.TestCheckResourceAttrPair("vkcs_kubernetes_cluster_v2.base", "network_id", "vkcs_networking_network.base", "id"),
+					resource.TestCheckResourceAttrPair("vkcs_kubernetes_cluster_v2.base", "subnet_id", "vkcs_networking_subnet.base", "id"),
+					resource.TestCheckResourceAttrPair("vkcs_kubernetes_cluster_v2.base", "loadbalancer_subnet_id", "vkcs_networking_subnet.base", "id"),
 
-					resource.TestCheckResourceAttrSet("vkcs_kubernetes_cluster_v2.basic", "id"),
-					resource.TestCheckResourceAttrSet("vkcs_kubernetes_cluster_v2.basic", "uuid"),
-					resource.TestCheckResourceAttrSet("vkcs_kubernetes_cluster_v2.basic", "external_network_id"),
-					resource.TestCheckResourceAttrSet("vkcs_kubernetes_cluster_v2.basic", "k8s_config"),
-					resource.TestCheckResourceAttrSet("vkcs_kubernetes_cluster_v2.basic", "created_at"),
-					resource.TestCheckResourceAttrSet("vkcs_kubernetes_cluster_v2.basic", "status"),
-					resource.TestCheckResourceAttrSet("vkcs_kubernetes_cluster_v2.basic", "project_id"),
-					resource.TestCheckResourceAttrSet("vkcs_kubernetes_cluster_v2.basic", "api_address"),
-					resource.TestCheckResourceAttrSet("vkcs_kubernetes_cluster_v2.basic", "api_lb_vip"),
-					resource.TestCheckResourceAttrSet("vkcs_kubernetes_cluster_v2.basic", "region"),
+					resource.TestCheckResourceAttrSet("vkcs_kubernetes_cluster_v2.base", "id"),
+					resource.TestCheckResourceAttrSet("vkcs_kubernetes_cluster_v2.base", "uuid"),
+					resource.TestCheckResourceAttrSet("vkcs_kubernetes_cluster_v2.base", "external_network_id"),
+					resource.TestCheckResourceAttrSet("vkcs_kubernetes_cluster_v2.base", "k8s_config"),
+					resource.TestCheckResourceAttrSet("vkcs_kubernetes_cluster_v2.base", "created_at"),
+					resource.TestCheckResourceAttrSet("vkcs_kubernetes_cluster_v2.base", "status"),
+					resource.TestCheckResourceAttrSet("vkcs_kubernetes_cluster_v2.base", "project_id"),
+					resource.TestCheckResourceAttrSet("vkcs_kubernetes_cluster_v2.base", "api_address"),
+					resource.TestCheckResourceAttrSet("vkcs_kubernetes_cluster_v2.base", "api_lb_vip"),
+					resource.TestCheckResourceAttrSet("vkcs_kubernetes_cluster_v2.base", "region"),
 				),
 			},
-			acctest.ImportStep("vkcs_kubernetes_cluster_v2.basic"),
+			acctest.ImportStep("vkcs_kubernetes_cluster_v2.base"),
 		},
 	})
 }
@@ -109,7 +110,7 @@ func TestAccKubernetesClusterV2_full(t *testing.T) {
 					resource.TestCheckResourceAttr("vkcs_kubernetes_cluster_v2.full", "node_groups.#", "0"),
 
 					resource.TestCheckResourceAttrPair("vkcs_kubernetes_cluster_v2.full", "external_network_id", "data.vkcs_networking_network.base-extnet", "id"),
-					resource.TestCheckResourceAttrPair("vkcs_kubernetes_cluster_v2.full", "master_flavor", "data.vkcs_compute_flavor.base", "id"),
+					resource.TestCheckResourceAttrPair("vkcs_kubernetes_cluster_v2.full", "master_flavor", "data.vkcs_compute_flavor.master", "id"),
 					resource.TestCheckResourceAttrPair("vkcs_kubernetes_cluster_v2.full", "network_id", "vkcs_networking_network.base", "id"),
 					resource.TestCheckResourceAttrPair("vkcs_kubernetes_cluster_v2.full", "subnet_id", "vkcs_networking_subnet.base", "id"),
 					resource.TestCheckResourceAttrPair("vkcs_kubernetes_cluster_v2.full", "loadbalancer_subnet_id", "vkcs_networking_subnet.base", "id"),
@@ -148,7 +149,7 @@ func TestAccKubernetesClusterV2_scale(t *testing.T) {
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckKubernetesClusterV2Exists("vkcs_kubernetes_cluster_v2.scale"),
 					resource.TestCheckResourceAttr("vkcs_kubernetes_cluster_v2.scale", "name", clusterName),
-					resource.TestCheckResourceAttrPair("vkcs_kubernetes_cluster_v2.scale", "master_flavor", "data.vkcs_compute_flavor.base", "id"),
+					resource.TestCheckResourceAttrPair("vkcs_kubernetes_cluster_v2.scale", "master_flavor", "data.vkcs_compute_flavor.master", "id"),
 				),
 			},
 			{
@@ -160,7 +161,7 @@ func TestAccKubernetesClusterV2_scale(t *testing.T) {
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckKubernetesClusterV2Exists("vkcs_kubernetes_cluster_v2.scale"),
 					resource.TestCheckResourceAttr("vkcs_kubernetes_cluster_v2.scale", "name", clusterName),
-					resource.TestCheckResourceAttrPair("vkcs_kubernetes_cluster_v2.scale", "master_flavor", "data.vkcs_compute_flavor.base", "id"),
+					resource.TestCheckResourceAttrPair("vkcs_kubernetes_cluster_v2.scale", "master_flavor", "data.vkcs_compute_flavor.master", "id"),
 				),
 			},
 		},
@@ -226,33 +227,32 @@ func testAccCheckKubernetesClusterV2Exists(n string) resource.TestCheckFunc {
 			return fmt.Errorf("error creating Kubernetes API client: %s", err)
 		}
 
-		found, err := clusters.Get(client, rs.Primary.ID).Extract()
+		_, err = clusters.Get(client, rs.Primary.ID).Extract()
 		if err != nil {
+			if errutil.IsNotFound(err) {
+				return errors.New("Kubernetes cluster not found")
+			}
 			return err
-		}
-
-		if found == nil {
-			return errors.New("Kubernetes cluster not found")
 		}
 
 		return nil
 	}
 }
 
-const testAccKubernetesClusterV2Basic = `
+const testAccKubernetesClusterV2Base = `
 {{ .TestAccKubernetesNetworkingBase }}
 
-data "vkcs_compute_flavor" "base" {
+data "vkcs_compute_flavor" "master" {
   name = "Standard-6-12"
 }
 
-resource "vkcs_kubernetes_cluster_v2" "basic" {
+resource "vkcs_kubernetes_cluster_v2" "base" {
   name                   = "{{ .ClusterName }}"
   version                = "v1.34.2"
   cluster_type           = "standard"
   availability_zones     = ["MS1"]
   master_count           = 1
-  master_flavor          = data.vkcs_compute_flavor.base.id
+  master_flavor          = data.vkcs_compute_flavor.master.id
   network_plugin         = "calico"
   pods_ipv4_cidr         = "10.100.0.0/16"
   network_id             = vkcs_networking_network.base.id
@@ -268,7 +268,7 @@ resource "vkcs_kubernetes_cluster_v2" "basic" {
 const testAccKubernetesClusterV2Full = `
 {{ .TestAccKubernetesNetworkingBase }}
 
-data "vkcs_compute_flavor" "base" {
+data "vkcs_compute_flavor" "master" {
   name = "Standard-6-12"
 }
 
@@ -280,7 +280,7 @@ resource "vkcs_kubernetes_cluster_v2" "full" {
   cluster_type           = "standard"
   availability_zones     = ["MS1"]
   master_count           = 1
-  master_flavor          = data.vkcs_compute_flavor.base.id
+  master_flavor          = data.vkcs_compute_flavor.master.id
   network_plugin         = "calico"
   pods_ipv4_cidr         = "10.100.0.0/16"
   network_id             = vkcs_networking_network.base.id
@@ -306,7 +306,7 @@ resource "vkcs_kubernetes_cluster_v2" "full" {
 const testAccKubernetesClusterV2Scale = `
 {{ .TestAccKubernetesNetworkingBase }}
 
-data "vkcs_compute_flavor" "base" {
+data "vkcs_compute_flavor" "master" {
   name = "{{ .FlavorName }}"
 }
 
@@ -316,7 +316,7 @@ resource "vkcs_kubernetes_cluster_v2" "scale" {
   cluster_type           = "standard"
   availability_zones     = ["MS1"]
   master_count           = 1
-  master_flavor          = data.vkcs_compute_flavor.base.id
+  master_flavor          = data.vkcs_compute_flavor.master.id
   network_plugin         = "calico"
   network_id             = vkcs_networking_network.base.id
   subnet_id              = vkcs_networking_subnet.base.id
@@ -341,7 +341,7 @@ resource "vkcs_kubernetes_cluster_v2" "scale" {
 const testAccKubernetesClusterV2Upgrade = `
 {{ .TestAccKubernetesNetworkingBase }}
 
-data "vkcs_compute_flavor" "base" {
+data "vkcs_compute_flavor" "master" {
   name = "Standard-6-12"
 }
 
@@ -351,7 +351,7 @@ resource "vkcs_kubernetes_cluster_v2" "upgrade" {
   cluster_type           = "standard"
   availability_zones     = ["MS1"]
   master_count           = 1
-  master_flavor          = data.vkcs_compute_flavor.base.id
+  master_flavor          = data.vkcs_compute_flavor.master.id
   network_plugin         = "calico"
   network_id             = vkcs_networking_network.base.id
   subnet_id              = vkcs_networking_subnet.base.id
@@ -374,7 +374,7 @@ resource "vkcs_kubernetes_cluster_v2" "upgrade" {
 resource "vkcs_kubernetes_node_group_v2" "upgrade" {
   cluster_id               = vkcs_kubernetes_cluster_v2.upgrade.id
   name                     = "tfacc-upgrade-v2"
-  node_flavor              = data.vkcs_compute_flavor.base.id
+  node_flavor              = data.vkcs_compute_flavor.master.id
   availability_zone        = "MS1"
   disk_type                = "ceph-ssd"
   disk_size                = 30